requirements.txt
djangorestframework==3.9.4
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-rest-swagger